web/requirements.txt
Django
gunicorn
Pillow
psycopg2-binary
django-allauth
django-crispy-forms
braintree
